Integrate Faces/Mojarra 5 snapshot with com.sun to org.glassfish
