+++ appserver/web/jsf-connector/src/main/resources/META-INF/services/com.sun.faces.spi.injectionprovider
@@ -1,1 +1,1 @@
-org.glassfish.faces.integration.GlassFishInjectionProvider:org.glassfish.faces.integration.GlassFishInjectionProvider
+org.glassfish.faces.integration.GlassFishHighAvailabilityEnabler:org.glassfish.faces.integration.GlassFishHighAvailabilityEnabler
