+++ nucleus/parent/pom.xml
@@ -194,8 +194,8 @@
         <jakarta.annotation-api.version>3.0.0</jakarta.annotation-api.version>
 
         <!-- Jakarta Faces -->
-        <jakarta.faces-api.version>5.0.0-M1</jakarta.faces-api.version>
-        <mojarra.version>5.0.0-M1</mojarra.version>
+        <jakarta.faces-api.version>5.0.0-SNAPSHOT</jakarta.faces-api.version>
+        <mojarra.version>5.0.0-SNAPSHOT</mojarra.version>
 
         <!-- Jakarta WebSocket -->
         <jakarta.websocket-api.version>2.3.0-M2</jakarta.websocket-api.version>
